format: stop relying on array properties
There are numerous libraries out there (sugarjs, prototypejs) that
extend the `Array` prototype. The `simplify` function in js-quantities
relies on setting/checking properties on an array instance, which means
that it's incompatible with such libraries. Add compatibility by using a
separate object for keeping track of the counters instead of using
properties on array instances.

Fix #130.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>